LayoutTests/imported/w3c/web-platform-tests/content-security-policy/meta/sandbox-iframe-expected.txt

self is derived correctly inside inside a sandboxed iframe.


FAIL img-src 'self' works when specified in a meta tag. assert_equals: expected "PASS" but got "FAIL"
